ENH: Add a new node to connect common information
Add a new node to the graph which is named by a particular date/time.
This node should be used to connect entries across the data types that
have the same date/time.  In this early case, we try to connect all other
data entries to the date/time of a visit.