ENH: Add timeline and connection highlighting
Add a time axis to the bottom of the visualization that spans the time
that is covered by the earliest and latest visits.

Change the "Visit" nodes to be constrained to this axis in y, while
being anchored at their appropriate place along the date scale.

Add a click event to the "Visit" nodes that highlights the links
that have that node as a source. It should turn the relevant links red and
make the link thicker.

Loosen the date matching constraints  for all patient information types.
Instead of matching the exact date/time for each item to a visit,
the connections should be made between visits and objects by only
matching the date.